vciso-backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
anthropic==0.75.0
pydantic==2.12.5
pydantic-settings==2.6.1
python-dotenv==1.2.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
